Basic file list tab added
